Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/addEventListener-expected.txt

CONSOLE MESSAGE: line 20: ReferenceError: Can't find variable: undefined_variable
CONSOLE MESSAGE: line 23: SyntaxError: Unexpected token ')'

PASS window.onerror - addEventListener 
